+++ funcx_endpoint/funcx_endpoint/engines/helper.py
@@ -9,9 +9,9 @@ from funcx_common.tasks import ActorName, TaskState
 
 from funcx.errors import MaxResultSizeExceeded
 from funcx.serialize import FuncXSerializer
+from funcx_endpoint.engines.high_throughput.messages import Message
 from funcx_endpoint.exception_handling import get_error_string, get_result_error_details
 from funcx_endpoint.exceptions import CouldNotExecuteUserTaskError
-from funcx_endpoint.executors.high_throughput.messages import Message
 
 log = logging.getLogger(__name__)
