+++ funcx_endpoint/funcx_endpoint/engines/high_throughput/funcx_manager.py
@@ -25,15 +25,15 @@ from funcx_common.tasks import ActorName, TaskState
 from parsl.version import VERSION as PARSL_VERSION
 
 from funcx.serialize import FuncXSerializer
-from funcx_endpoint.exception_handling import get_error_string, get_result_error_details
-from funcx_endpoint.executors.high_throughput.container_sched import naive_scheduler
-from funcx_endpoint.executors.high_throughput.mac_safe_queue import mpQueue
-from funcx_endpoint.executors.high_throughput.messages import (
+from funcx_endpoint.engines.high_throughput.container_sched import naive_scheduler
+from funcx_endpoint.engines.high_throughput.mac_safe_queue import mpQueue
+from funcx_endpoint.engines.high_throughput.messages import (
     ManagerStatusReport,
     Message,
     Task,
 )
-from funcx_endpoint.executors.high_throughput.worker_map import WorkerMap
+from funcx_endpoint.engines.high_throughput.worker_map import WorkerMap
+from funcx_endpoint.exception_handling import get_error_string, get_result_error_details
 from funcx_endpoint.logging_config import FXLogger, setup_logging
 
 RESULT_TAG = 10
